Move some functions out of views + fix a typo
