+++ projektrouska/functions.py
@@ -58,3 +58,10 @@ def calcmd5(string):
 
     # printing the equivalent hexadecimal value.
     return result.hexdigest()
+
+# source: https://stackoverflow.com/questions/8906926/formatting-timedelta-objects
+def strfdelta(tdelta, fmt):
+    d = {"days": tdelta.days}
+    d["hours"], rem = divmod(tdelta.seconds, 3600)
+    d["minutes"], d["seconds"] = divmod(rem, 60)
+    return fmt.format(**d)
